Accessing OR replacing sspm_dataset model elements

All methods described here allow to access the elements of contained in objects of class sspm_dataset.

spm_data(sspm_object)

# S4 method for class 'sspm_dataset'
spm_data(sspm_object)

spm_data(object) <- value

# S4 method for class 'sspm_dataset'
spm_data(object) <- value

# S4 method for class 'sspm_dataset'
spm_name(sspm_object)

# S4 method for class 'sspm_dataset'
spm_name(object) <- value

# S4 method for class 'sspm_dataset'
spm_unique_ID(sspm_object)

# S4 method for class 'sspm_dataset'
spm_unique_ID(object) <- value

spm_coords_col(sspm_object)

# S4 method for class 'sspm_dataset'
spm_coords_col(sspm_object)

spm_coords_col(object) <- value

# S4 method for class 'sspm_dataset'
spm_coords_col(object) <- value

# S4 method for class 'sspm_dataset'
spm_time(sspm_object)

# S4 method for class 'sspm_dataset'
spm_time(object) <- value

spm_biomass_vars(sspm_object)

# S4 method for class 'sspm_dataset'
spm_biomass_vars(sspm_object)

spm_biomass_vars(object) <- value

# S4 method for class 'sspm_dataset'
spm_biomass_vars(object) <- value

spm_density_vars(sspm_object)

# S4 method for class 'sspm_dataset'
spm_density_vars(sspm_object)

spm_density_vars(object) <- value

# S4 method for class 'sspm_dataset'
spm_density_vars(object) <- value

spm_formulas(sspm_object)

# S4 method for class 'sspm_dataset'
spm_formulas(sspm_object)

spm_formulas(object) <- value

# S4 method for class 'sspm_dataset'
spm_formulas(object) <- value

# S4 method for class 'sspm_dataset'
spm_smoothed_data(sspm_object)

# S4 method for class 'sspm_dataset'
spm_smoothed_data(object) <- value

spm_smoothed_fit(sspm_object)

# S4 method for class 'sspm_dataset'
spm_smoothed_fit(sspm_object)

spm_smoothed_fit(object) <- value

# S4 method for class 'sspm_dataset'
spm_smoothed_fit(object) <- value

spm_smoothed_vars(sspm_object)

# S4 method for class 'sspm_dataset'
spm_smoothed_vars(sspm_object)

spm_smoothed_vars(object) <- value

# S4 method for class 'sspm_dataset'
spm_smoothed_vars(object) <- value

is_mapped(sspm_object)

# S4 method for class 'sspm_dataset'
is_mapped(sspm_object)

is_mapped(object) <- value

# S4 method for class 'sspm_dataset'
is_mapped(object) <- value

# S4 method for class 'sspm_dataset'
spm_boundaries(sspm_object)

# S4 method for class 'sspm_dataset'
spm_boundaries(object) <- value

Arguments

sspm_object

[sspm_dataset] An object of class sspm_dataset.

object

[sspm_dataset] An object of class sspm_dataset.

value

typically an array-like R object of a similar class as x.

Value

The object in the required slot.

Examples

data(borealis_simulated, package = "sspm")
biomass_dataset <- spm_as_dataset(data.frame(borealis_simulated), name = "borealis",
                                  density = "weight_per_km2",
                                  time = "year_f",
                                  coords = c('lon_dec','lat_dec'),
                                  uniqueID = "uniqueID")
#> ℹ  Casting data matrix into simple feature collection using columns: lon_dec, lat_dec
#> !  Warning: sspm is assuming WGS 84 CRS is to be used for casting
spm_data(biomass_dataset)
#> Simple feature collection with 1800 features and 8 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -64.13859 ymin: 46.01421 xmax: -46.86366 ymax: 60.81292
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>    year_f sfa      weight_per_km2 temp_at_bottom   lon_dec  lat_dec row
#> 1    1995   7  14.70372 [kg/km^2]      1.3054569 -62.82016 59.19084   1
#> 2    1995   7 365.72370 [kg/km^2]      0.0000000 -56.70318 55.19117   2
#> 3    1995   7 108.39527 [kg/km^2]      0.9818056 -60.60656 56.88572   3
#> 4    1995   7  62.31766 [kg/km^2]      0.0000000 -57.98659 56.42020   4
#> 5    1995   7   0.00000 [kg/km^2]      0.6982752 -59.04377 57.08357   5
#> 6    1995   7   0.00000 [kg/km^2]      1.4510729 -58.79227 56.30274   6
#> 7    1995   7 236.15051 [kg/km^2]      1.0511097 -51.05171 50.69749   7
#> 8    1995   7  68.19999 [kg/km^2]      0.0000000 -54.06121 54.52411   8
#> 9    1995   7  93.62156 [kg/km^2]      1.8644427 -54.94126 53.67627   9
#> 10   1995   7  25.42231 [kg/km^2]      1.7439115 -54.71695 50.45287  10
#>      uniqueID                   geometry
#> 1   y1995s7r1 POINT (-62.82016 59.19084)
#> 2   y1995s7r2 POINT (-56.70318 55.19117)
#> 3   y1995s7r3 POINT (-60.60656 56.88572)
#> 4   y1995s7r4  POINT (-57.98659 56.4202)
#> 5   y1995s7r5 POINT (-59.04377 57.08357)
#> 6   y1995s7r6 POINT (-58.79227 56.30274)
#> 7   y1995s7r7 POINT (-51.05171 50.69749)
#> 8   y1995s7r8 POINT (-54.06121 54.52411)
#> 9   y1995s7r9 POINT (-54.94126 53.67627)
#> 10 y1995s7r10 POINT (-54.71695 50.45287)