Accessing OR replacing sspm_fit model elements

All methods described here allow to access the elements of contained in objects of class sspm_fit.

# S4 method for class 'sspm_fit'
spm_unique_ID(sspm_object)

# S4 method for class 'sspm_fit'
spm_unique_ID(object) <- value

# S4 method for class 'sspm_fit'
spm_time(sspm_object)

# S4 method for class 'sspm_fit'
spm_time(object) <- value

# S4 method for class 'sspm_fit'
spm_formulas(sspm_object)

# S4 method for class 'sspm_fit'
spm_formulas(object) <- value

# S4 method for class 'sspm_fit'
spm_smoothed_data(sspm_object)

# S4 method for class 'sspm_fit'
spm_smoothed_data(object) <- value

spm_get_fit(sspm_object)

# S4 method for class 'sspm_fit'
spm_get_fit(sspm_object)

spm_get_fit(object) <- value

# S4 method for class 'sspm_fit'
spm_get_fit(object) <- value

# S4 method for class 'sspm_fit'
spm_boundaries(sspm_object)

# S4 method for class 'sspm_fit'
spm_boundaries(object) <- value

# S4 method for class 'sspm_fit'
spm_boundary(sspm_object)

# S4 method for class 'sspm_fit'
spm_boundary(object) <- value

Arguments

sspm_object

[sspm_fit] An object of class sspm_fit.

object

[sspm_fit] An object of class sspm_fit.

value

typically an array-like R object of a similar class as x.

Value

The object in the required slot.

Examples

if (FALSE) { # \dontrun{
spm_formulas(sspm_object)
...
} # }