Map dataset onto a discretized sspm object

After discretizing, the next step is to map coavriate datasets.

map_dataset(sspm_object, data, ...)

# S4 method for sspm_discrete,data.frame
map_dataset(sspm_object, data, ...)

# S4 method for sspm_discrete,sf
map_dataset(sspm_object, data, ...)

# S4 method for sspm_discrete,sspm_data
map_dataset(sspm_object, data, ...)

# S4 method for sspm_discrete,list
map_dataset(sspm_object, data, ...)

# S4 method for sspm,ANY
map_dataset(sspm_object, data, ...)

Arguments

sspm_object
[sspm_discrete] An object of class sspm_discrete.

data
[data.frame OR sf] The dataset.

...
Arguments passed on to as_sspm_data

time_col

[character] The column of data for the temporal dimensions (i.e. year).

coords

[character] The column of data for longitude and latitude of the observations.

name

[character] The name of the dataset, default to "Biomass".

uniqueID

[character] The column of data that is unique for all rows of the data matrix.

crs

Coordinate reference system, passed onto st_as_sf.

Value

The updated object, of class sspm_discrete.

Functions

map_dataset,sspm_discrete,data.frame-method: TODO

map_dataset,sspm_discrete,sf-method: TODO

map_dataset,sspm_discrete,sspm_data-method: TODO

map_dataset,sspm_discrete,list-method: TODO

map_dataset,sspm,ANY-method: TODO