sspm Smoothing functions

A full sspm formula contains calls to the smoothing terms smooth_time(), smooth_space(), smooth_space_time().

smooth_time(
  type = "ICAR",
  dataset,
  sspm_object,
  k = NULL,
  bs = "re",
  xt = NULL,
  ...
)

smooth_space(
  type = "ICAR",
  dataset,
  sspm_object,
  k = 30,
  bs = "mrf",
  xt = NULL,
  ...
)

smooth_space_time(
  type = "ICAR",
  dataset,
  sspm_object,
  k = NULL,
  bs = c("re", "mrf"),
  xt = NULL,
  ...
)

# S4 method for ANY,character,sspm_discrete
smooth_time(
  type = "ICAR",
  dataset,
  sspm_object,
  k = NULL,
  bs = "re",
  xt = NULL,
  ...
)

# S4 method for ANY,character,sspm_discrete
smooth_space(
  type = "ICAR",
  dataset,
  sspm_object,
  k = 30,
  bs = "mrf",
  xt = NULL,
  ...
)

# S4 method for ANY,character,sspm_discrete
smooth_space_time(
  type = "ICAR",
  dataset,
  sspm_object,
  k = NULL,
  bs = c("re", "mrf"),
  xt = NULL,
  ...
)

Arguments

type
[character] Type of smooth, currently only "ICAR" is supported.

dataset
[character] The name of the dataset among base and/or mapped datasets for which to specify the formula

sspm_object
[sspm_discrete] An object of class sspm_discrete.

k
the dimension of the basis used to represent the smooth term. The default depends on the number of variables that the smooth is a function of. k should not be less than the dimension of the null space of the penalty for the term (see null.space.dimension), but will be reset if it is. See choose.k for further information.

bs
a two letter character string indicating the (penalized) smoothing basis to use. (eg "tp" for thin plate regression spline, "cr" for cubic regression spline). see smooth.terms for an over view of what is available.

xt
Any extra information required to set up a particular basis. Used e.g. to set large data set handling behaviour for "tp" basis.

...
Further arguments passed down, none used for now.

Value

A list of 2 lists:

args, contains the arguments to be passed on to the mgcv smooths

vars, contains variables relevant to the evaluation of the smooth.

Functions

smooth_space: TODO

smooth_space_time: TODO

smooth_time,ANY,character,sspm_discrete-method: TODO

smooth_space,ANY,character,sspm_discrete-method: TODO

smooth_space_time,ANY,character,sspm_discrete-method: TODO