Accessing OR replacing sspm model elements

All methods described here allow to access the elements of contained in objects of the different classes of the package.

spm_name(sspm_object)

spm_name(object) <- value

spm_datasets(sspm_object)

# S4 method for class 'sspm'
spm_datasets(sspm_object)

spm_datasets(object) <- value

# S4 method for class 'sspm'
spm_datasets(object) <- value

spm_boundaries(sspm_object)

# S4 method for class 'sspm'
spm_boundaries(sspm_object)

spm_boundaries(object) <- value

# S4 method for class 'sspm'
spm_boundaries(object) <- value

spm_smoothed_data(sspm_object)

# S4 method for class 'sspm'
spm_smoothed_data(sspm_object)

spm_smoothed_data(object) <- value

# S4 method for class 'sspm'
spm_smoothed_data(object) <- value

spm_time(sspm_object)

# S4 method for class 'sspm'
spm_time(sspm_object)

spm_time(object) <- value

# S4 method for class 'sspm'
spm_time(object) <- value

is_split(sspm_object)

# S4 method for class 'sspm'
is_split(sspm_object)

is_split(object) <- value

# S4 method for class 'sspm'
is_split(object) <- value

spm_unique_ID(sspm_object)

# S4 method for class 'sspm'
spm_unique_ID(sspm_object)

spm_unique_ID(object) <- value

# S4 method for class 'sspm'
spm_unique_ID(object) <- value

Arguments

sspm_object

[sspm OR adjacent] An object of class sspm or others derivative classes.

object

[sspm OR adjacent] An object of class sspm or others derivative classes.

value

typically an array-like R object of a similar class as x.

Value

The object in the required slot.

Examples

data(borealis_simulated, package = "sspm")
biomass_dataset <- spm_as_dataset(data.frame(borealis_simulated), name = "borealis",
                                  density = "weight_per_km2",
                                  time = "year_f",
                                  coords = c('lon_dec','lat_dec'),
                                  uniqueID = "uniqueID")
#> ℹ  Casting data matrix into simple feature collection using columns: lon_dec, lat_dec
#> !  Warning: sspm is assuming WGS 84 CRS is to be used for casting
spm_name(biomass_dataset)
#> [1] "borealis"