sspm: Spatial Surplus Production Model Framework for Northern Shrimp Populations

The goal of `sspm` is to implement a gam-based spatial surplus production model, aimed at modeling northern shrimp population in Canada but potentially to any stock in any location. The package is opinionated in its implementation of SPMs as it internally makes the choice to use penalized spatial gams with time lags based on Pedersen et al. (2020). However, it also aims to provide options for the user to customize their model.

See also

Author

Maintainer: Valentin Lucet valentin.lucet@gmail.com

Authors:

Eric Pedersen eric.pedersen@concordia.ca