sspm boundary structure

One of the first step in the sspm workflow is to create one or more object(s) of class sspm_boundary from an sf object.

Functions

sspm_discrete_boundary-class: sspm_discrete_boundary

Slots

boundaries

[sf] Spatial boundaries (polygons).

boundary_column

[character] The column of data that represents the spatial boundaries.

surface_column

[character] The column of data that represents the area of spatial boundaries.

method

[discretization_method] (if discrete) discretization method used.

patches

[sf] (if discrete) Patches resulting from discretization.

points

[sf] (if discrete) Sample points used for discretization.