sspm dataset structure

One of the first step in the sspm workflow is to create one or more object(s) of class sspm_dataset from a data.frame, tibble or sf object.

Slots

name

[character] The name of the dataset, default to "Biomass".

data

[data.frame OR sf OR tibble] The dataset.

biomass

[character] The biomass columns of data.

density

[character] The biomass density columns of data.

time_column

[character] The column of data that represents the temporal dimension of the dataset.

coords

[character] The columns of data that represent the spatial dimension of the dataset: the two columns for longitude and latitude of the observations.

uniqueID

[character] The column of data that is unique for all rows of the data matrix.

boundaries

[sspm_discrete_boundary] Spatial boundaries (polygons).

formulas

[list] List of sspm_formula objects that specifies the smoothed variables.

smoothed_data

[ANY (sf)] The smoothed data.

smoothed_vars

[character] A vector storing the smoothed vars.

smoothed_fit

[list] The fit from smoothing the data

is_mapped

[logical] Whether the dataset has been mapped to boundaries (used internally).