Discretize a sspm model object

Discretize a sspm model object with a function from a discretization_method object class.

spm_discretize(sspm_object, with_dataset, discretization_method, ...)

# S4 method for ANY,ANY,missingOrNULL
spm_discretize(sspm_object, with_dataset, discretization_method, ...)

# S4 method for ANY,missing,ANY
spm_discretize(sspm_object, with_dataset, discretization_method, ...)

# S4 method for sspm,character,discretization_method
spm_discretize(sspm_object, with_dataset, discretization_method, ...)

# S4 method for sspm,character,character
spm_discretize(sspm_object, with_dataset, discretization_method, ...)

# S4 method for sspm_discrete,character,character
spm_discretize(sspm_object, with_dataset, discretization_method, ...)

# S4 method for sspm_discrete,ANY,ANY
spm_discretize(sspm_object, with_dataset, discretization_method, ...)

Arguments

sspm_object
[sspm] An object of class sspm.

with_dataset
[character] The name of the dataset to use for the discretization.

discretization_method
[character OR discretization_method] Either a character from the list of available methods (see spm_methods for the list) OR an object of class discretization_method.

...
[named list] Further arguments to be passed onto the function used in the discretization_method.

Value

An object of class sspm_discrete (the updated and discretized sspm object given as input).

Functions

spm_discretize,ANY,ANY,missingOrNULL-method: TODO

spm_discretize,ANY,missing,ANY-method: TODO

spm_discretize,sspm,character,discretization_method-method: TODO

spm_discretize,sspm,character,character-method: TODO

spm_discretize,sspm_discrete,character,character-method: TODO

spm_discretize,sspm_discrete,ANY,ANY-method: TODO